Split cloudaccount methods (#25)
